Extract duplicated createMockFetch into shared test-utils module
