+++ packages/miot-calendar-client/src/__tests__/client.test.ts
@@ -1,26 +1,7 @@
 import { describe, it, expect } from "vitest";
 import { createMiotCalendarClient, MiotCalendarApiError } from "../index.js";
 import type { ErrorResponse } from "../types.js";
-
-function createMockFetch(response: unknown, status = 200) {
-  const call = { url: "", init: {} as RequestInit };
-  const fn = async (url: string | URL | Request, init: RequestInit = {}) => {
-    if (typeof url === "string") {
-      call.url = url;
-    } else if (url instanceof URL) {
-      call.url = url.href;
-    } else {
-      call.url = url.url;
-    }
-    call.init = init;
-    return {
-      ok: status >= 200 && status < 300,
-      status,
-      json: async () => response,
-    } as Response;
-  };
-  return { fn, call };
-}
+import { createMockFetch } from "./test-utils.js";
 
 describe("createMiotCalendarClient", () => {
   const baseUrl = "https://api.example.com";
