+++ packages/miot-calendar-client/src/__tests__/test-utils.ts
@@ -1,0 +1,19 @@
+export function createMockFetch<T>(response: T, status = 200) {
+  const call = { url: "", init: {} as RequestInit };
+  const fn = async (url: string | URL | Request, init: RequestInit = {}) => {
+    if (typeof url === "string") {
+      call.url = url;
+    } else if (url instanceof URL) {
+      call.url = url.href;
+    } else {
+      call.url = url.url;
+    }
+    call.init = init;
+    return {
+      ok: status >= 200 && status < 300,
+      status,
+      json: async () => response,
+    } as Response;
+  };
+  return { fn, call };
+}
